Basic working Boggle
