+++ client/boggle/index.html
@@ -1,0 +1,143 @@
+<head>
+    <style>
+        .square {
+            display: flex;
+            justify-content: center;
+            vertical-align: middle;
+            align-items: center;
+        }
+        .square:after {
+            content: "";
+            display: block;
+            padding-bottom: 100%;
+        }
+
+        .selected-letter {
+            border: 2px solid #fff;
+            background-color: #333;
+            color: #fff;
+        }
+    </style>
+</head>
+<body style="padding: 5%;">
+    <div class="container">
+        <div id="row0" class="row"><div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div></div>
+        <div id="row1" class="row"><div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div></div>
+        <div id="row2" class="row"><div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div></div>
+        <div id="row3" class="row"><div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div></div>
+        <div id="scoreboard" class="row" style="padding-top: 5%">
+            <div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div>
+            <div class="col-xs-6 col-sm-4 col-md-2 col-lg-2 btn btn-default" onclick="submit()">Iontráil focal</div>
+            <div class="col-xs-6 col-sm-4 col-md-2 col-lg-2 btn btn-default" onclick="reset()">Glan an bord</div>
+        </div>
+        <div class="row">
+            <div class="col-xs-0 col-sm-2 col-md-4 col-lg-4"></div>
+            <div class="col-xs-12 col-sm-8 offset-sm-2 col-md-4 offset-md-4 col-lg-4 offset-lg-4">
+                <h3 id="score">Scór: 0</h3>
+                <h3>Aimsithe:</h3>
+                <div id="found"></div>
+            </div>
+        </div>
+    </div>
+    <script>
+        let selected = -1;
+        let letters = "";
+        let wordsFound = [];
+        let score = 0;
+
+        for (let i = 0; i < 16 ; i++) {
+            let div = document.createElement('div');
+            let content = document.createElement('h3');
+            div.className = "col-xs-3 col-sm-2 col-md-1 col-lg-1 square";
+            div.style.borderRadius = "10%";
+            div.style.border = "2px solid #555";
+            div.id = i.toString();
+            content.innerText = i.toString();
+            div.appendChild(content);
+            document.getElementById("row" + Math.floor(i / 4)).appendChild(div);
+            div.setAttribute("onclick", "select(" + i + ")");
+        }
+
+        function select(i) {
+            if (isConnected(i)) {
+               selected = i;
+               document.getElementById(i.toString()).className += " selected-letter";
+               letters += document.getElementById(i.toString()).children[0].innerText.toString();
+            }
+        }
+
+        function reset(){
+            selected = -1;
+            letters = "";
+            for (let i = 0; i < 16 ; i++) {
+                document.getElementById(i.toString()).className = "col-xs-3 col-sm-2 col-md-1 col-lg-1 square";
+            }
+        }
+
+        function getScore(i) {
+            if (i < 3) return 0;
+            switch(i) {
+                case 3: return 1;
+                case 4: return 2;
+                case 5: return 3;
+                case 6: return 4;
+                case 7: return 5;
+                default: return 6;
+            }
+        }
+
+        function submit(){
+            if (wordsFound.indexOf(letters) === -1) {
+                if (focail.indexOf(letters) !== -1) {
+                    score += getScore(letters.length);
+                    wordsFound.push(letters);
+                    document.getElementById("score").innerText = "Scór: " + score;
+                    document.getElementById("found").innerText = wordsFound.join(", ");
+                } else alert("Ní focail í '" + letters + "'");
+            } else alert("D'aimsigh tú an focal '" + letters + "' cheana");
+            reset();
+        }
+
+        function isConnected(i) {
+            let normalVals = [1,3,4,5];
+            let leftSideOOB = [-1, -5, 3];
+            let rightSideOOB = [1, 5, -3];
+            if (selected === -1) return true;
+            else if (i % 4 === 0 && leftSideOOB.indexOf(selected - i) !== -1) return false;
+            else if (i % 4 === 3 && rightSideOOB.indexOf(selected - i) !== -1) return false;
+            else return normalVals.indexOf(Math.abs(selected - i)) !== -1;
+        }
+
+        let charMap = {"a": 69486, "i": 54990, "h": 41582, "c": 34750, "r": 34237, "t": 31328, "e": 27948, "n": 27821, "l": 26597, "s": 25468, "o": 22497, "m": 16642, "d": 13815, "í": 11916, "g": 11219, "á": 10929, "b": 9458, "p": 8366, "ó": 8096, "é": 7614, "f": 7154, "ú": 7093, "u": 6755, "v": 1080, "z": 557, "k": 356, "y": 243, "w": 173, "j": 161, "x": 148, "q": 92};
+        let charMapSize = calculateMapSize();
+
+        function calculateMapSize() {
+            let x = 0;
+            for (k in charMap) x += charMap[k];
+            return x;
+        }
+
+        function generateBoard() {
+            for (let i = 0; i < 16 ; i++) {
+                document.getElementById(i.toString()).children[0].innerText = getChar();
+            }
+        }
+
+        function getChar() {
+            let rand = Math.floor(Math.random() * charMapSize);
+            for (k in charMap) {
+                // console.log(rand);
+                if (rand <= charMap[k]) return k;
+                else rand -= charMap[k];
+            }
+            return "a";
+        }
+
+        generateBoard();
+    </script>
+</body>
+
+<!-- Bootstrap CDN -->
+<link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
+
+<script src="/client/boggle/assets/focal.txt"></script>
