+++ client/index.html
@@ -1,9 +1,12 @@
 
 <body style = "padding-left: 5%; padding-top: 5%">
-<div class="col-sm-12 col-md-12 col-lg-12">
+<div class=" col-xs-12 col-sm-12 col-md-12 col-lg-12">
     <h2>Fáilte go cluichí as Gaeilge!</h2>
-    <div class="col-sm-12 col-md-12 col-lg-12 btn btn-default" onclick="location.href='/cartai';">
-        Cártaí i gCoinne Daonnachta
+    <div class="col-xs-12 col-sm-12 col-md-4 col-lg-3 btn btn-default" onclick="location.href='/cartai';">
+        Cártaí i gCoinne Daonnachta (briste)
+    </div>
+    <div class="col-xs-12 col-sm-12 col-md-4 col-lg-3 btn btn-default" onclick="location.href='/boggle';">
+        Boggle
     </div>
 </div>
